feature-policy/picture-in-picture-supported-by-feature-policy.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('picture-in-picture', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise picture-in-picture.');
</script>
